feat: implement ResultOverlay component for win/lose feedback with confetti animation
