--- src/components/elements/Victory/VictoryOverlayComponent.scss
@@ -1,82 +1,0 @@
-.win-overlay {
-	position: fixed;
-	inset: 0;
-	z-index: 9999;
-	display: grid;
-	grid-template-rows: auto 1fr;
-	pointer-events: none;
-	padding: 0 1rem;
-}
-
-.win-overlay::before {
-	content: '';
-}
-
-.win-card {
-	justify-self: center;
-	align-self: start;
-	margin-top: 1rem;
-	max-width: 480px;
-	width: min(92vw, 480px);
-	pointer-events: auto;
-	background: var(--bg);
-	color: var(--fg);
-	border: 1px solid var(--border2, color-mix(in oklch, var(--fg) 20%, var(--bg) 80%));
-	border-radius: var(--card-radius, 12px);
-	box-shadow:
-		0 10px 30px var(--fg/15),
-		0 2px 6px var(--fg/20);
-	padding: 1rem 1.25rem;
-	text-align: center;
-	animation:
-		toastSlideDown 260ms cubic-bezier(0.2, 0.8, 0.2, 1),
-		fadeIn 220ms ease-out;
-}
-
-.win-emoji {
-	font-size: 1.75rem;
-	line-height: 1;
-	margin-bottom: 0.25rem;
-}
-
-.win-card h2 {
-	margin: 0.25rem 0 0.25rem;
-	font-size: 1.25rem;
-	font-weight: 700;
-}
-
-.win-card p {
-	margin: 0 0 0.75rem;
-	opacity: 0.9;
-	font-size: 0.95rem;
-}
-
-.win-actions {
-	display: inline-flex;
-	gap: 0.5rem;
-	flex-wrap: wrap;
-	justify-content: center;
-}
-
-@keyframes fadeIn {
-	from {
-		opacity: 0;
-	}
-	to {
-		opacity: 1;
-	}
-}
-@keyframes toastSlideDown {
-	from {
-		transform: translateY(-8px);
-	}
-	to {
-		transform: translateY(0);
-	}
-}
-
-@media (prefers-reduced-motion: reduce) {
-	.win-card {
-		animation: none;
-	}
-}
